--- core/src/main/java/com/unstage/core/jobposting/entity/JobPosting.java
@@ -1,31 +1,0 @@
-package com.unstage.core.jobposting.entity;
-
-import com.unstage.core.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "job_postings")
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Builder
-public class JobPosting extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String title;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "welfare_center_id", nullable = false,
-            foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
-    private WelfareCenter welfareCenter;
-
-    private LocalDateTime recruitmentStartDate;
-    private LocalDateTime recruitmentEndDate;
-}
