--- core/src/main/java/com/unstage/core/jobposting/repository/JobPostingRepository.java
@@ -1,9 +1,0 @@
-package com.unstage.core.jobposting.repository;
-
-import com.unstage.core.jobposting.entity.JobPosting;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface JobPostingRepository extends JpaRepository<JobPosting, Long> {
-}
