--- core/src/main/java/com/unstage/core/post/entity/Category.java
@@ -1,16 +1,0 @@
-package com.unstage.core.post.entity;
-
-public enum Category {
-    NOTICE("공지사항"),
-    JOB("채용");
-
-    private final String description;
-
-    Category(String description) {
-        this.description = description;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-}
