--- core/src/main/java/com/unstage/core/post/entity/Post.java
@@ -1,45 +1,0 @@
-package com.unstage.core.post.entity;
-
-import com.unstage.core.BaseEntity;
-import com.unstage.core.jobposting.entity.WelfareCenter;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name = "posts")
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Builder
-public class Post extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String title;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "welfare_center_id", nullable = false,
-            foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
-    private WelfareCenter welfareCenter;
-
-    private String url;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Category category;
-
-    @Builder.Default
-    @OneToMany(mappedBy = "post", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<PostJob> postJobs = new ArrayList<>();
-
-    public void addPostJob(PostJob postJob) {
-        this.postJobs.add(postJob);
-        postJob.setPost(this);
-    }
-}
