--- core/src/main/java/com/unstage/core/post/entity/PostJob.java
@@ -1,29 +1,0 @@
-package com.unstage.core.post.entity;
-
-import com.unstage.core.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "post_jobs")
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Builder
-public class PostJob extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private LocalDateTime recruitmentStartDate;
-    private LocalDateTime recruitmentEndDate;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id", nullable = false,
-            foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
-    @Setter
-    private Post post;
-}
