--- core/src/main/java/com/unstage/core/post/service/PostService.java
@@ -1,27 +1,0 @@
-package com.unstage.core.post.service;
-
-import com.unstage.core.paging.PageParams;
-import com.unstage.core.paging.PageResponse;
-import com.unstage.core.post.component.PostReader;
-import com.unstage.core.post.dto.GetPostsResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class PostService {
-
-    private final PostReader postReader;
-
-    public PageResponse<GetPostsResponse> getPostings(final PageParams pageParams) {
-        return postReader.readAllWithPage(pageParams.page(), pageParams.size());
-    }
-
-    public PageResponse<GetPostsResponse> getNoticePosts(int page, int size) {
-        return postReader.readNoticeWithPage(page, size);
-    }
-
-    public PageResponse<GetPostsResponse> getJobPosts(int page, int size) {
-        return postReader.readJobWithPage(page, size);
-    }
-}
