+++ core/src/main/java/com/unstage/core/welfarecenter/entity/PostJob.java
@@ -1,0 +1,19 @@
+package com.unstage.core.welfarecenter.entity;
+
+import jakarta.persistence.DiscriminatorValue;
+import jakarta.persistence.Entity;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Entity
+@DiscriminatorValue("job")
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Builder
+public class PostJob extends Post {
+
+    private LocalDateTime recruitmentStartDate;
+    private LocalDateTime recruitmentEndDate;
+}
