+++ core/src/main/java/com/unstage/core/welfarecenter/entity/PostNotice.java
@@ -1,0 +1,14 @@
+package com.unstage.core.welfarecenter.entity;
+
+import jakarta.persistence.DiscriminatorValue;
+import jakarta.persistence.Entity;
+import lombok.Getter;
+
+@Entity
+@DiscriminatorValue("notice")
+@Getter
+public class PostNotice extends Post {
+    public PostNotice() {
+    }
+}
+
